:recycle: Refactor how callbacks are created
